fix: refactor defintions after suggestions by @peterblazejewicz
